Fixed CMakeLists.txt and compilation warnings

+++ easy/chuck.norris/main.cpp
@@ -19,9 +19,9 @@ int main()
     
     std::string result;
     int last_char = -1;
-    for (int i = 0; i < input.size(); i++)
+    for (size_t i = 0; i < input.size(); i++)
     {
-        for (int j = 0; j < 7; j++)
+        for (size_t j = 0; j < 7; j++)
         {
             int bit = input[i] & (0x01 << (6 - j)) ? 1 : 0;
             if (-1 == last_char
